First Commit
Date: 1/5/24

Interactive analysis report showing the effects of success on finding success in stable housing. The poster was what was presented to the Mayor of Waterville, ME as well as the board of the Mid-Maine Homeless Shelter. The application shown as "app.R" was the interactive analysis report created from our findings.